requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.0; python_version >= '3.7'
click==8.0.3
dj-database-url==0.5.0
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-filter==21.1
django-heroku==0.3.1
django==4.0.2
djangorestframework==3.13.1
gunicorn==20.1.0
pillow==9.0.0
psycopg2==2.9.3; python_version >= '3.6'
python-dotenv[cli]==0.19.2
pytz==2021.3
sqlparse==0.4.2; python_version >= '3.5'
whitenoise==5.3.0; python_version >= '3.5' and python_version < '4'
